exception en empleado

+++ empleado-service/src/main/java/com/barbershop/empleado_service/service/EmpleadoService.java
@@ -1,7 +1,9 @@
 package com.barbershop.empleado_service.service;
 
+import com.barbershop.empleado_service.exception.ResourceNotFoundException;
 import com.barbershop.empleado_service.model.Empleado;
 import com.barbershop.empleado_service.repository.EmpleadoRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,7 +20,11 @@ public class EmpleadoService {
     }
 
     public Empleado findById(Long id) {
-        return repository.findById(id).orElse(null);
+
+        return repository.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException(
+                                "Empleado con ID " + id + " no encontrado"));
     }
 
     public Empleado save(Empleado empleado) {
@@ -26,6 +32,12 @@ public class EmpleadoService {
     }
 
     public void delete(Long id) {
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Empleado con ID " + id + " no encontrado");
+        }
+
         repository.deleteById(id);
     }
